Extract methods for determining if a node is suitable to be written on a single line to `CheckSingleLineSuitability`.
